requirements.txt
torch>=2.0.0
transformers>=4.37.0
pdf2image>=1.16.3
Pillow>=10.0.0 